Edit structure of project, use live data every where.Add mock Data for content view and prepare for testing add Localizable
