+++ QuantiSpaceX.xcodeproj/xcuserdata/petrzacek.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -10,5 +10,13 @@
 			<integer>0</integer>
 		</dict>
 	</dict>
+	<key>SuppressBuildableAutocreation</key>
+	<dict>
+		<key>70092BF62D1E065500774A37</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+	</dict>
 </dict>
 </plist>
