+++ QuantiSpaceX/Localizable/Localizable.xcstrings
@@ -1,0 +1,80 @@
+{
+  "sourceLanguage" : "en",
+  "strings" : {
+    "%@ second%@ burn time" : {
+      "localizations" : {
+        "cs" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "%1$@ sekund čas spalování"
+          }
+        },
+        "en" : {
+          "stringUnit" : {
+            "state" : "new",
+            "value" : "%1$@ second%2$@ burn time"
+          }
+        }
+      }
+    },
+    "%@ ton%@ of fuel" : {
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "new",
+            "value" : "%1$@ ton%2$@ of fuel"
+          }
+        }
+      }
+    },
+    "%lld engine%@" : {
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "new",
+            "value" : "%1$lld engine%2$@"
+          }
+        }
+      }
+    },
+    "first flight: %@" : {
+
+    },
+    "Launch" : {
+      "localizations" : {
+        "cs" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Start"
+          }
+        }
+      }
+    },
+    "Launch succesful!" : {
+      "localizations" : {
+        "cs" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Start byl úspěšný!"
+          }
+        }
+      }
+    },
+    "Move your phone up\nto launch the rocket" : {
+
+    },
+    "Overview" : {
+
+    },
+    "Parameters" : {
+
+    },
+    "Photos" : {
+
+    },
+    "Rockets" : {
+
+    }
+  },
+  "version" : "1.0"
+}
