+++ QuantiSpaceX/RocketList/RocketListData/CompositeFairing.swift
@@ -1,0 +1,13 @@
+//
+//  CompositeFairing.swift
+//  QuantiSpaceX
+//
+//  Created by Petr Žáček on 14.02.2025.
+//
+
+import Foundation
+
+struct CompositeFairing: Codable, Hashable {
+    var height: Height
+    var diameter: Diameter
+}
