+++ QuantiSpaceX/RocketList/RocketListData/Diameter.swift
@@ -1,0 +1,14 @@
+//
+//  Diameter.swift
+//  QuantiSpaceX
+//
+//  Created by Petr Žáček on 14.02.2025.
+//
+
+import Foundation
+
+struct Diameter: Codable, Hashable {
+    var meters: Double?
+    var feet: Double?
+}
+
