+++ QuantiSpaceX/RocketList/RocketListData/Height.swift
@@ -1,0 +1,13 @@
+//
+//  Height.swift
+//  QuantiSpaceX
+//
+//  Created by Petr Žáček on 14.02.2025.
+//
+
+import Foundation
+
+struct Height: Codable, Hashable {
+    var meters: Double?
+    var feet: Double?
+}
