+++ QuantiSpaceX/RocketList/RocketListData/Isp.swift
@@ -1,0 +1,13 @@
+//
+//  Isp.swift
+//  QuantiSpaceX
+//
+//  Created by Petr Žáček on 14.02.2025.
+//
+
+import Foundation
+
+struct Isp: Codable, Hashable {
+    var sea_level: Double
+    var vacuum: Double
+}
