+++ QuantiSpaceX/RocketList/RocketListData/LandingLegs.swift
@@ -1,0 +1,13 @@
+//
+//  LandingLegs 2.swift
+//  QuantiSpaceX
+//
+//  Created by Petr Žáček on 14.02.2025.
+//
+
+import Foundation
+
+struct LandingLegs: Codable, Hashable {
+    var number: Int?
+    var material: String?
+}
