+++ QuantiSpaceX/RocketList/RocketListData/Mass.swift
@@ -1,0 +1,13 @@
+//
+//  Mass.swift
+//  QuantiSpaceX
+//
+//  Created by Petr Žáček on 14.02.2025.
+//
+
+import Foundation
+
+struct Mass: Codable, Hashable {
+    var kg: Double?
+    var lb: Double?
+}
