+++ QuantiSpaceX/RocketList/RocketListData/Thrust.swift
@@ -1,0 +1,13 @@
+//
+//  Thrust.swift
+//  QuantiSpaceX
+//
+//  Created by Petr Žáček on 14.02.2025.
+//
+
+import Foundation
+
+struct Thrust: Codable, Hashable{
+    var kN: Double?
+    var lbf: Double?
+}
