+++ QuantiSpaceX/RocketList/RocketListData/ThrustSeaLevel.swift
@@ -1,0 +1,13 @@
+//
+//  ThrustSeaLevel.swift
+//  QuantiSpaceX
+//
+//  Created by Petr Žáček on 14.02.2025.
+//
+
+import Foundation
+
+struct ThrustSeaLevel: Codable, Hashable {
+    var kn: Double?
+    var lbf: Double?
+}
