+++ QuantiSpaceX/RocketList/RocketListData/ThrustVacuum.swift
@@ -1,0 +1,13 @@
+//
+//  ThrustVacuum.swift
+//  QuantiSpaceX
+//
+//  Created by Petr Žáček on 14.02.2025.
+//
+
+import Foundation
+
+struct ThrustVacuum: Codable, Hashable {
+    var kn: Double?
+    var lbf: Double?
+}
